feat:implemented the functionality for renaming the file
