+++ src/utils/commandProcessing.js
@@ -1,8 +1,9 @@
-import { cd } from "../commands/cd.js";
-import { ls } from "../commands/ls.js";
+import { cd } from "../commands/navigation/cd.js";
+import { ls } from "../commands/navigation/ls.js";
+import { rename } from "../commands/fs/rename.js";
 
 export const commandProcessing = (command, params) => {
-  const commands = { cd, ls };
+  const commands = { cd, ls, rename };
 
   if (commands[command]) {
